Fix circular dependency between data-broker and indicators
