+++ packages/data-broker/src/utils.ts
@@ -3,7 +3,9 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { Analytic, AnalyticCluster, ArrayAnalytic } from './types';
+import { Analytic } from './types';
+
+type ArrayAnalytic = [string, string, string, string | number];
 
 const DATA_ELEMENT = 'dataElement';
 const DATA_GROUP = 'dataGroup';
@@ -12,23 +14,6 @@ export const DATA_SOURCE_TYPES = {
   DATA_ELEMENT,
   DATA_GROUP,
   SYNC_GROUP,
-};
-
-/**
- * Groups { dataElement: value } pairs per organisationUnit/period combination
- */
-export const analyticsToAnalyticClusters = (analytics: Analytic[]): AnalyticCluster[] => {
-  const clusterMap: Record<string, AnalyticCluster> = {};
-  analytics.forEach(analytic => {
-    const { dataElement, organisationUnit, period, value } = analytic;
-    const key = `${organisationUnit}__${period}`;
-    if (!clusterMap[key]) {
-      clusterMap[key] = { organisationUnit, period, dataValues: {} };
-    }
-    clusterMap[key].dataValues[dataElement] = value;
-  });
-
-  return Object.values(clusterMap);
 };
 
 export const arrayToAnalytics = (arrayAnalytics: ArrayAnalytic[]): Analytic[] =>
