+++ packages/indicators/src/__tests__/Builder/AnalyticArithmeticBuilder/AnalyticArithmeticBuilder.fixtures.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { arrayToAnalytics } from '@tupaia/data-broker';
+import { arrayToAnalytics } from '../../../utils';
 import { Analytic } from '../../../types';
 
 /**
